added all comments and refactored viewmodel into one class
